requirements.txt
# Core framework
streamlit==1.28.1
python-dotenv==1.0.0
pandas==2.1.3
plotly==5.17.0
requests==2.31.0
beautifulsoup4==4.12.2

# LangChain ecosystem
langchain==0.0.350
langchain-core==0.3.75
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-groq==0.0.3   # ✅ Added Groq support

# Other utilities
typing-extensions>=4.5.0
